Move components to their own folder
