+++ src/components/InputFile.js
@@ -1,8 +1,8 @@
 /* @flow */
 import * as React from 'react';
 
-import type { Rows, Parser } from './index.js.flow';
-import parseFile from './parseFile';
+import type { Rows, Parser } from '../index.js.flow';
+import parseFile from '../parseFile';
 
 type InputFileProps = {
     inputRef?: (input: ?HTMLInputElement) => void,
