+++ src/components/index.js
@@ -1,0 +1,5 @@
+/* @flow */
+import DropZone from './DropZone';
+import InputFile from './InputFile';
+
+export { DropZone, InputFile };
